pyproject.toml
[project]
name = "recoleta"
version = "0.1.0"
description = "Personal research intelligence funnel for ingesting, analyzing, and publishing signals."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "anthropic>=0.46.0",
    "apscheduler>=3.11.2",
    "arxiv>=2.4.0",
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.2",
    "litellm>=1.76.1",
    "loguru>=0.7.3",
    "markdownify>=1.2.2",
    "marker-pdf>=1.10.2",
    "openai>=1.109.1",
    "openreview-py>=1.57.1",
    "orjson>=3.11.7",
    "platformdirs>=4.9.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-slugify>=8.0.4",
    "python-telegram-bot>=22.6",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.3",
    "rich>=14.3.3",
    "sqlmodel>=0.0.37",
    "tenacity>=9.1.4",
    "trafilatura>=2.0.0",
    "typer>=0.24.1",
]

[project.scripts]
recoleta = "recoleta.cli:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
